improve

+++ controllers/calendar.controller.js
@@ -43,7 +43,7 @@ export const getCalendar = async (req, res, next) => {
 
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return next(errorHandler(400, 'Validation failed', errors.array()));
+        return next(errorHandler(400, 'Validation fails when trying to get calendar', errors.array()));
     }
     const { id } = matchedData(req, { locations: ['params'] });
 
@@ -51,10 +51,14 @@ export const getCalendar = async (req, res, next) => {
         const calendar = await Calendar.findOne({ _id: id });
 
         if (!calendar) {
-            throw new Error('Calendar not found');
+            throw errorHandler(404, 'Calendar not found when trying to fetch Calendar');
         }
 
-        res.status(200).json({ calendar });
+        if (calendar.userID.toString() !== req.user.id) {
+            throw errorHandler(403, 'Unauthorized Access when trying to fetch Calendar');
+        }
+
+        res.status(200).json({ message: 'Calendar fetched successfully', data: calendar });
     } catch (error) {
         next(error);
     }
@@ -63,7 +67,7 @@ export const getCalendar = async (req, res, next) => {
 export const deleteCalendar = async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return next(errorHandler(400, 'Validation failed', errors.array()));
+        return next(errorHandler(400, 'Validation fails when trying to delete calendar', errors.array()));
     }
     const { id } = matchedData(req, { locations: ['params'] });
 
@@ -71,10 +75,23 @@ export const deleteCalendar = async (req, res, next) => {
 
     try {
         session.startTransaction();
+
+        const calendar = await Calendar.findOne({
+            _id: id,
+        });
+
+        if (!calendar) {
+            throw errorHandler(404, 'Calendar not found when trying to delete Calendar');
+        }
+
+        if (calendar.userID.toString() !== req.user.id) {
+            throw errorHandler(403, 'Unauthorized Access when trying to delete Calendar');
+        }
+
         const deleteResult = await Calendar.deleteOne({ _id: id }, { session });
 
         if (!deleteResult.acknowledged || deleteResult.deletedCount !== 1) {
-            throw new Error('Calendar deletion failed');
+            throw errorHandler(500, 'Calendar deletion failed when trying to delete Calendar');
         }
 
         const updateResult = await User.updateOne(
@@ -84,7 +101,7 @@ export const deleteCalendar = async (req, res, next) => {
         );
 
         if (!updateResult.acknowledged || updateResult.modifiedCount !== 1 || updateResult.matchedCount !== 1) {
-            throw new Error('Calendar deletion failed');
+            throw errorHandler(500, 'An error occurred while updating the user\'s calendars section.');
         }
 
         await session.commitTransaction();
@@ -101,7 +118,7 @@ export const deleteCalendar = async (req, res, next) => {
 export const updateCalendar = async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return next(errorHandler(400, 'Validation failed', errors.array()));
+        return next(errorHandler(400, 'Validation fails when trying to update calendar', errors.array()));
     }
 
     const { id } = matchedData(req, { locations: ['params'] });
@@ -114,11 +131,11 @@ export const updateCalendar = async (req, res, next) => {
         const calendar = await Calendar.findOne({ _id: id });
 
         if (!calendar) {
-            throw new Error('Calendar not found');
+            throw errorHandler(404, 'Calendar not found when trying to update Calendar');
         }
 
         if (calendar.userID.toString() !== req.user.id) {
-            throw new Error('You can update only your calendar!');
+            throw errorHandler(403, 'Unauthorized Access when trying to update Calendar');
         }
 
         const updateResult = await Calendar.updateOne(
@@ -128,7 +145,7 @@ export const updateCalendar = async (req, res, next) => {
         );
 
         if (!updateResult.acknowledged || updateResult.modifiedCount !== 1 || updateResult.matchedCount !== 1) {
-            throw new Error('Calendar update failed');
+            throw errorHandler(500, 'Calendar update failed');
         }
 
         const updateResultUser = await User.updateOne(
@@ -138,7 +155,7 @@ export const updateCalendar = async (req, res, next) => {
         );
 
         if (!updateResultUser.acknowledged || updateResultUser.modifiedCount !== 1 || updateResultUser.matchedCount !== 1) {
-            throw new Error('Calendar update failed');
+            throw errorHandler(500, 'An error occurred while updating the user\'s calendars section.');
         }
 
         await session.commitTransaction();
@@ -165,7 +182,7 @@ export const createSubCalendar = async (req, res, next) => {
         const calendar = await Calendar.findOne({ _id: id });
 
         if (!calendar) {
-            throw new Error('Calendar not found');
+            throw errorHandler(404, 'Calendar not found');
         }
 
         const createdSubCalendar = await SubCalendar.create({
